% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc.R
\name{bin_plot}
\alias{bin_plot}
\title{bin_plot}
\usage{
bin_plot(x, y, bins, threads = 1L)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{bins}{bin num}

\item{threads}{thread num}
}
\description{
Plot y mean based on x bins
}
