% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_utils.R
\name{color.alpha}
\alias{color.alpha}
\title{color.alpha}
\usage{
color.alpha(col, alpha = 0.2)
}
\arguments{
\item{col}{str col}

\item{alpha}{alpha transparency, where 1 means fully opaque and 0 fully transparent, default = 0.2}
}
\description{
Functions for calculating transparent and desaturated colors.
}
