% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_utils.R
\name{color.desat}
\alias{color.desat}
\title{color.desat}
\usage{
color.desat(acol, amt = 0.5)
}
\arguments{
\item{amt}{amount of desaturation of color to apply, where 1 means totally desaturated (grayscale), default = 0.5}

\item{col}{str col}
}
\description{
Functions for calculating transparent and desaturated colors.
}
