% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_utils.R
\name{color.dist}
\alias{color.dist}
\title{color.dist}
\usage{
color.dist(x, mu = 0, sd = 1, col = "slateblue")
}
\arguments{
\item{x}{A vector of values to use for calculating distances}

\item{mu}{default = 0.}

\item{sd}{default = 1.}

\item{col}{default = slateblue}
}
\description{
Functions for calculating transparent and desaturated colors.
}
