% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc.R
\name{diverge_plot}
\alias{diverge_plot}
\title{diverge_plot}
\usage{
diverge_plot(dt, factor_col, value_col, threshold)
}
\arguments{
\item{dt}{data.table}

\item{factor_col}{str factor col}

\item{value_col}{str value col}

\item{threshold}{threshold to divide two group}
}
\description{
plot value_col based on factor_col, and label above/below tag for given threshold
to_plot_dt <- dt[, tail(.SD, 1L), by = factor_col][order(value_col)]
}
