% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc.R
\name{hist_plot}
\alias{hist_plot}
\title{hist_plot}
\usage{
hist_plot(x, bins, min_val = NA, max_val = NA, threads = 1L)
}
\arguments{
\item{x}{vector}

\item{bins}{bin num}

\item{min_val}{min val}

\item{max_val}{max val}

\item{threads}{thread num}
}
\description{
Plot hist
}
