% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_plot.R
\name{mat_plot}
\alias{mat_plot}
\title{mat_plot}
\usage{
mat_plot(dt, first_n = 0L, last_n = 0L, xlab = NULL, ylab = NULL, main = NULL)
}
\arguments{
\item{dt}{data}

\item{first_n}{0L first n columns to highlight}

\item{last_n}{0L last n columns to highlight}
}
\description{
Functions for plot table
}
