% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_plot.R
\name{plot_bar}
\alias{plot_bar}
\title{plot_bar}
\usage{
plot_bar(dt, category_col, value_col, facet_col = NA)
}
\arguments{
\item{dt}{data}

\item{category_col}{to display}

\item{value_col}{to display}

\item{facet_col}{group by}
}
\description{
Functions for plot bar
}
