% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_plot.R
\name{plot_heat}
\alias{plot_heat}
\title{plot_heat}
\usage{
plot_heat(dt, x_col, y_col, value_col)
}
\arguments{
\item{dt}{data}

\item{x_col}{to group}

\item{y_col}{to group}

\item{value_col}{to display}
}
\description{
Functions for plot heat
}
