% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_plot.R
\name{plot_hist}
\alias{plot_hist}
\title{plot_hist}
\usage{
plot_hist(
  dt,
  value_col,
  facet_col = NA,
  n_bins = 10L,
  use_density_hight = TRUE
)
}
\arguments{
\item{dt}{data}

\item{value_col}{to display}

\item{facet_col}{group by}

\item{n_bins}{default to 10 bins}

\item{use_density_hight}{show density}
}
\description{
Functions for plot histgram
}
