% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_plot.R
\name{plot_qq}
\alias{plot_qq}
\title{plot_qq}
\usage{
plot_qq(dt, value_col, facet_col = NA)
}
\arguments{
\item{dt}{data}

\item{value_col}{to display}

\item{facet_col}{group by}
}
\description{
Functions for plot qq
}
