% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc.R
\name{plot_quantile}
\alias{plot_quantile}
\title{plot_quantile}
\usage{
plot_quantile(
  data,
  x,
  y,
  w = 1,
  n = 100,
  na.rm = TRUE,
  ...,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{data}{dt}

\item{x}{vec}

\item{y}{vec}

\item{w}{weight vec}

\item{na.rm}{remove missing values}

\item{...}{additional parameters passed to plot}

\item{xlab}{x label}

\item{ylab}{y label}
}
\description{
Plot with quantile sampling
}
