% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_plot.R
\name{plot_scatter}
\alias{plot_scatter}
\title{plot_scatter}
\usage{
plot_scatter(dt, x_col, y_col, facet_col = NA, color_col = NA)
}
\arguments{
\item{dt}{data}

\item{x_col}{to display}

\item{y_col}{to display}

\item{facet_col}{group by}

\item{color_col}{group by}
}
\description{
Functions for plot scatter
}
