% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_plot.R
\name{plot_violin}
\alias{plot_violin}
\title{plot_violin}
\usage{
plot_violin(dt, x_col = NA, y_col, facet_col = NA, color_col = NA)
}
\arguments{
\item{dt}{data}

\item{x_col}{to display}

\item{y_col}{to display}

\item{facet_col}{group by}

\item{color_col}{group by}
}
\description{
Functions for plot violin
}
